requirements-mobile.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diskcache==5.6.3
faiss-cpu==1.10.0
filelock==3.18.0
flask==3.1.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
llama-cpp-python==0.2.24
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pillow==11.2.1
pymupdf==1.25.5
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
zipp==3.21.0
